requirements.txt
pandas
snowflake-connector-python
streamlit
langchain
langchain_community
openai
python-dotenv